+++ src/unit-test-coverage/ut-stubs/src/bsp-console-impl-stubs.c
@@ -43,7 +43,7 @@
 
        Note: This should write the string as-is without buffering.
  ------------------------------------------------------------------*/
-void OCS_OS_BSP_ConsoleOutput_Impl(const char *Str, uint32_t DataLen)
+void OCS_OS_BSP_ConsoleOutput_Impl(const char *Str, size_t DataLen)
 {
     int32_t retcode = UT_DEFAULT_IMPL(OCS_OS_BSP_ConsoleOutput_Impl);
